APEX-RAG-PDF-Chatbot/requirements.txt
gradio==4.16.0
langchain==0.1.0
langchain-groq==0.0.1
langchain-core==0.1.10
langchain-community==0.0.13
langchain-text-splitters==0.0.1
pypdf==3.17.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pyrebase4==4.7.1
python-dotenv==1.0.0
